+++ common/src/main/java/dev/xylonity/tooltipoverhaul/client/util/ColorUtils.java
@@ -1,6 +1,5 @@
 package dev.xylonity.tooltipoverhaul.client.util;
 
-import dev.xylonity.tooltipoverhaul.client.old.Palette;
 import dev.xylonity.tooltipoverhaul.client.old.frame.CustomFrameData;
 import dev.xylonity.tooltipoverhaul.client.render.TooltipContext;
 import dev.xylonity.tooltipoverhaul.config.TooltipsConfig;