+++ common/src/main/java/dev/xylonity/tooltipoverhaul/client/util/Constants.java
@@ -3,7 +3,7 @@ package dev.xylonity.tooltipoverhaul.client.util;
 public final class Constants {
 
     public static int ICON_SIZE = 22;
-    public static int ICON_SIZE_PADDING = 12;
+    public static int ICON_SIZE_PADDING = ICON_SIZE - 10;
     public static int SEPARATION_TITLE_ICON = 3;
     public static int DIVIDER_LINE_FULL_PADDING = 6;